EdgeFunctions/DataSimulation/modules/DataSimulator/requirements.txt
azure-iot-device~=2.7.0
faker